+++ src/main/java/ru/mail/polis/dao/dogm/NoSuchElementExceptionLite.java
@@ -1,4 +1,4 @@
-package ru.mail.polis.dao;
+package ru.mail.polis.dao.dogm;
 
 import java.util.NoSuchElementException;
